requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
markdown2==2.5.3
werkzeug==3.1.3
python-dotenv==1.1.0
